write itemset_comparison.py
